+++ cups/dest.c
@@ -572,15 +572,20 @@ cupsConnectDest(
     if (cb)
       (*cb)(user_data, CUPS_DEST_FLAGS_UNCONNECTED | CUPS_DEST_FLAGS_CONNECTING, dest);
 
-    if (httpReconnect(http, msec, cancel) && cb)
+    if (!httpReconnect(http, msec, cancel) && cb)
     {
       if (cancel && *cancel)
 	(*cb)(user_data, CUPS_DEST_FLAGS_UNCONNECTED | CUPS_DEST_FLAGS_CONNECTING, dest);
       else
 	(*cb)(user_data, CUPS_DEST_FLAGS_UNCONNECTED | CUPS_DEST_FLAGS_ERROR, dest);
+
+      httpClose(http);
+      http = NULL;
     }
     else if (cb)
+    {
       (*cb)(user_data, CUPS_DEST_FLAGS_NONE, dest);
+    }
   }
 
   return (http);